+++ Game/mapwindow.hh
@@ -27,7 +27,7 @@ public:
     ~MapWindow();
 
     void setSize(int width, int height);
-    void setScale(double scale);
+    void setScale(int scale);
     void resize();
     void drawItem( std::shared_ptr<Course::GameObject> obj);
     void removeItem( std::shared_ptr<Course::GameObject> obj);
